# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dbaff4c29bdd95510990c0e
  title: TestowyProjekt
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /vindicationcase/:
    post:
      x-skaffolder-id: 5dbb027729bdd95510990d9b
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dbb018229bdd95510990d74
      x-skaffolder-resource: VindicationCase
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - VindicationCase
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VindicationCase"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ActualDebt:
                  type: number
                AddDate:
                  type: integer
                StartingDebt:
                  type: number
    get:
      x-skaffolder-id: 5dbb027729bdd95510990da2
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dbb018229bdd95510990d74
      x-skaffolder-resource: VindicationCase
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - VindicationCase
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VindicationCase"
  /vindicationcase/{id}:
    delete:
      x-skaffolder-id: 5dbb027729bdd95510990da1
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dbb018229bdd95510990d74
      x-skaffolder-resource: VindicationCase
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - VindicationCase
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dbb027729bdd95510990d99
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dbb018229bdd95510990d74
      x-skaffolder-resource: VindicationCase
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - VindicationCase
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VindicationCase"
    post:
      x-skaffolder-id: 5dbb027729bdd95510990d9c
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dbb018229bdd95510990d74
      x-skaffolder-resource: VindicationCase
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - VindicationCase
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VindicationCase"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VindicationCase"
  /vindicationcase/findByDebtor/{key}:
    get:
      x-skaffolder-id: 5dbb027d29bdd95510990daa
      x-skaffolder-name: findByDebtor
      x-skaffolder-id-resource: 5dbb018229bdd95510990d74
      x-skaffolder-resource: VindicationCase
      x-skaffolder-crudAction: findByDebtor
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByDebtor
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByDebtor/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByDebtor
      tags:
        - VindicationCase
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa Debtor da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VindicationCase"
  /vindicationcase/findByDocument/{key}:
    get:
      x-skaffolder-id: 5dbb028229bdd95510990db5
      x-skaffolder-name: findByDocument
      x-skaffolder-id-resource: 5dbb018229bdd95510990d74
      x-skaffolder-resource: VindicationCase
      x-skaffolder-crudAction: findByDocument
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByDocument
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByDocument/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByDocument
      tags:
        - VindicationCase
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa Document da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VindicationCase"
    
    
# JWT security
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    VindicationCase:
      x-skaffolder-id: 5dbb018229bdd95510990d74
      x-skaffolder-id-db: 5dbaff4c29bdd95510990c0f
      x-skaffolder-id-entity: 5dbb018229bdd95510990d6f
      x-skaffolder-url: /vindicationcase
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        ActualDebt:
          type: number
          x-skaffolder-id-attr: 5dbb018229bdd95510990d71
          x-skaffolder-type: Decimal
        AddDate:
          type: integer
          x-skaffolder-id-attr: 5dbb018229bdd95510990d70
          x-skaffolder-type: Date
        StartingDebt:
          type: number
          x-skaffolder-id-attr: 5dbb018229bdd95510990d72
          x-skaffolder-type: Decimal
      x-skaffolder-relations: 
        Debtor:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5dbb018229bdd95510990d6f
          x-skaffolder-ent2: 5dbb01cd29bdd95510990d7e
        Document:
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5dbb018229bdd95510990d6f
          x-skaffolder-ent2: 5dbb025529bdd95510990d8c
      required:
        - _id
      
    
    
  # Databases
...