# src/main/java/com/testowyprojekt/controller/base/VindicationCaseBaseController.java
/* 
* Generated by
* 
*      _____ _          __  __      _     _
*     / ____| |        / _|/ _|    | |   | |
*    | (___ | | ____ _| |_| |_ ___ | | __| | ___ _ __
*     \___ \| |/ / _` |  _|  _/ _ \| |/ _` |/ _ \ '__|
*     ____) |   < (_| | | | || (_) | | (_| |  __/ |
*    |_____/|_|\_\__,_|_| |_| \___/|_|\__,_|\___|_|
*
* The code generator that works in many programming languages
*
*			https://www.skaffolder.com
*
*
* You can generate the code from the command-line
*       https://npmjs.com/package/skaffolder-cli
*
*       npm install -g skaffodler-cli
*
*   *   *   *   *   *   *   *   *   *   *   *   *   *   *   *
*
* To remove this comment please upgrade your plan here: 
*      https://app.skaffolder.com/#!/upgrade
*
* Or get up to 70% discount sharing your unique link:
*       https://app.skaffolder.com/#!/register?friend=5dbafeb029bdd95510990bea
*
* You will get 10% discount for each one of your friends
* 
*/
package com.testowyprojekt.controller.base;

import java.util.List;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.http.ResponseEntity;
import org.modelmapper.ModelMapper;
import java.util.stream.Collectors;
import java.util.ArrayList;
import java.util.Map;
import java.util.HashMap;
import org.springframework.security.access.annotation.Secured;
import org.springframework.beans.factory.annotation.Autowired;
import com.testowyprojekt.db.testowyprojekt_db.service.VindicationCaseService;
import com.testowyprojekt.db.testowyprojekt_db.entity.VindicationCase;
import com.testowyprojekt.db.testowyprojekt_db.dtos.VindicationCaseDto;

import com.testowyprojekt.db.testowyprojekt_db.entity.Debtor;

import com.testowyprojekt.db.testowyprojekt_db.entity.Document;

//IMPORT RELATIONS

import com.testowyprojekt.db.testowyprojekt_db.service.DebtorService;

import com.testowyprojekt.db.testowyprojekt_db.service.DocumentService;


public class VindicationCaseBaseController {
    
    @Autowired
	VindicationCaseService vindicationcaseService;

	
	@Autowired
	DebtorService debtorService;
	
	@Autowired
	DocumentService documentService;
	

	@Autowired
	private ModelMapper modelMapper;



//CRUD METHODS


    //CRUD - CREATE
    @Secured({ "ROLE_PRIVATE_USER" })
	@PostMapping("/vindicationcase")
	public ResponseEntity<VindicationCaseDto> insert(@RequestBody VindicationCase obj) {
				
		//external relation Debtor
		if (obj.getDebtor() != null) {
			if (obj.getDebtor().get_id() == null)
				debtorService.insert(obj.getDebtor());
			else
				debtorService.insert(debtorService.getOne(obj.getDebtor().get_id()));
		}
				
		//external relation Document
		if (obj.getDocument() != null && !obj.getDocument().isEmpty()) {
			obj.getDocument().forEach(document -> {
				document = documentService.getOne(document.get_id());
				document.addVindicationCase(obj);
				documentService.insert(document);
			});
		}
		
		
		return ResponseEntity.ok().body(convertToDto(vindicationcaseService.insert(obj)));
	}

	
    //CRUD - REMOVE
    @Secured({ "ROLE_PRIVATE_USER" })
	@DeleteMapping("/vindicationcase/{id}")
	public ResponseEntity<Void> delete(@PathVariable("id") Long id) {
		VindicationCase vindicationcaseSelected = vindicationcaseService.getOne(id);
		
		
		//external relation Document
		if (vindicationcaseSelected.getDocument() != null) {
			vindicationcaseSelected.getDocument().forEach(document -> {
				document.getVindicationCase().removeIf(vindicationcase -> vindicationcase.get_id() == vindicationcaseSelected.get_id());
				documentService.insert(document);
			});
		}
		
		vindicationcaseService.delete(id);
		return ResponseEntity.ok().build();
	}
	
    //CRUD - FIND BY Debtor
    
	@GetMapping("/vindicationcase/findByDebtor/{key}")
	public ResponseEntity<List<VindicationCaseDto>> findByDebtor(@PathVariable("key") Long id) {
		
		Debtor debtor = debtorService.getOne(id);
		List<VindicationCase> vindicationcases = vindicationcaseService.findByDebtor(debtor);
		List<VindicationCaseDto> vindicationcasesDto = vindicationcases.stream()
				.map(vindicationcase -> convertToDto(vindicationcase))
				.collect(Collectors.toList());
		
		
		
		

		return ResponseEntity.ok().body(vindicationcasesDto);
	}
    //CRUD - FIND BY Document
    
	@GetMapping("/vindicationcase/findByDocument/{key}")
	public ResponseEntity<List<VindicationCaseDto>> findByDocument(@PathVariable("key") Long id) {
		
		
		
		
		Document document = documentService.getOne(id);
		List<VindicationCase> vindicationcases = vindicationcaseService.findByDocument(document);
		List<VindicationCaseDto> vindicationcasesDto = vindicationcases.stream()
				.map(vindicationcase -> convertToDto(vindicationcase))
				.collect(Collectors.toList());
		

		return ResponseEntity.ok().body(vindicationcasesDto);
	}
	
    //CRUD - GET ONE
    @Secured({ "ROLE_PRIVATE_USER" })
	@GetMapping("/vindicationcase/{id}")
	public ResponseEntity<VindicationCaseDto> get(@PathVariable("id") Long id) {
		VindicationCase vindicationcaseSelected = vindicationcaseService.getOne(id);
		return ResponseEntity.ok().body(convertToDto(vindicationcaseSelected));
	}
	
	
    //CRUD - GET LIST
    @Secured({ "ROLE_PRIVATE_USER" })
	@GetMapping("/vindicationcase")
	public ResponseEntity<List<VindicationCaseDto>> getList() {
		List<VindicationCase> list = vindicationcaseService.getAll();
		List<VindicationCaseDto> listDto = list.stream()
				.map(vindicationcase -> convertToDto(vindicationcase))
				.collect(Collectors.toList());
		return ResponseEntity.ok().body(listDto);
	}


    //CRUD - EDIT
    @Secured({ "ROLE_PRIVATE_USER" })
	@PostMapping("/vindicationcase/{id}")
	public ResponseEntity<VindicationCaseDto> update(@RequestBody VindicationCase obj, @PathVariable("id") Long id) {
	    
		//external relation Debtor
		if (obj.getDebtor() != null) {
			if (obj.getDebtor().get_id() == null)
				debtorService.insert(obj.getDebtor());
			else
				debtorService.insert(debtorService.getOne(obj.getDebtor().get_id()));
		}
				
		//external relation Document
		if (obj.getDocument() != null) {
			obj.getDocument().forEach(document -> {
				document = documentService.getOne(document.get_id());
				document.addVindicationCase(obj);
				documentService.insert(document);
			});
		}
		
		
		return ResponseEntity.ok().body(convertToDto(vindicationcaseService.insert(obj)));
	}
	


/*
 * CUSTOM SERVICES
 * 
 *	These services will be overwritten and implemented in  Custom.js
 */


	private VindicationCaseDto convertToDto(VindicationCase vindicationcase) {
		VindicationCaseDto vindicationcaseDto = modelMapper.map(vindicationcase, VindicationCaseDto.class);
		return vindicationcaseDto;
	}
}